Refactor : Refactor pages and component's files and folders
