+++ src/components/index.js
@@ -1,7 +1,5 @@
-export { Sidebar } from './template/sidebar/Sidebar'
-export {Home} from './pages/home/Home'
-export {Card} from './template/Card'
-export {CardCategory} from './template/CardCategory'
-export {Navbar} from './template/Navbar'
-export {Footer} from './template/footer/Footer'
-export {Categories} from './pages/categories/Categories'
+export { Sidebar } from './sidebar/Sidebar'
+export {Card} from './card/Card'
+export {CardCategory} from './card/CardCategory'
+export {Navbar} from './navbar/Navbar'
+export {Footer} from './footer/Footer'
