+++ src/pages/categories/categories.css
