+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export { Home } from "./home/Home";
+export { Categories } from "./categories/Categories";
